+++ src/Form/WalkType.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace App\Form;
+
+use App\Entity\Animal;
+use App\Entity\Walk;
+use Symfony\Bridge\Doctrine\Form\Type\EntityType;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class WalkType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('start', null, [
+                'widget' => 'single_text',
+            ])
+            ->add('end', null, [
+                'widget' => 'single_text',
+            ])
+            ->add('animal', EntityType::class, [
+                'class' => Animal::class,
+                'choice_label' => 'id',
+            ])
+            ->add('submit', SubmitType::class);
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => Walk::class,
+        ]);
+    }
+}